Moved 4 failings tests to new test classes
