+++ pom.xml
@@ -111,6 +111,9 @@
 					<includes>
 						<include>**/*Test*</include>
 					</includes>
+                    <excludes>
+                        <exclude>**/*BrowserOnlyTest*</exclude>
+                    </excludes>
 					<parallel>methods</parallel>
 					<threadCount>4</threadCount>
 					<argLine>${failsafeArgLine}</argLine>
